1、  其他人接彪的时候，很奇怪系统自动产生了一个蒋有志接的彪，内容和eric发的是一致的。（参看群图） 2、多人应标互动，发飙人在一个互动窗口发的内容，会串到其他人的窗口共享。 3、一旦发飙人回复了，就变成自己应自己的飙了，在接飙里面也有，点击进去就是自己发的话。 4、新建彪里面的旁听提示，“一旦设置不可修改”，可去掉。 5、进行中的项目，“编辑彪”后，金额和时间现在还能修改，改金额后系统根据改的金额转账，必须改。 6、接完别人的彪，输入完毕，如果有个确认按钮就很好，现在只能回退，如果点错了点了x还会丢失。 7、测试过程中，短信提醒都没有了，测试时间十点，是有防骚扰限制吗？ 8、中彪人，在“我接的彪”里面应该可以打开看，没中标的打不开，现在是都打不开。中标人应该也不能旁听付费，没中标需要旁听付费。 9、付款旁听后，第一瞬间跳出来的是自己应彪的内容，退出后再次进来看到的就是中标者的内容。（群图） 10、结束后的彪，发飙人分享转发，5元彪还是显示1.5，不是3元，发飙人应该没有旁听按钮。 11、搜索界面还是可以旁听自己的彪。 12、搜索第一次正常，但没法二次搜索，显示乱码，约里面是正常的，可以多次搜索。 13、旁听大厅和历史大厅里面显示的星星还是0，参考“招募合伙人”彪。 14、后台人工增加“约的数量”。

+++ src/main/java/com/timeshare/domain/Bid.java
@@ -16,7 +16,7 @@ public class Bid extends BaseDomain{
     private String canAudit;
     private int clickRate;
     private int submitCount;
-    private BigDecimal score;
+    private int score;
     private String bidCatalog;
     private String stopReason;
 
@@ -65,11 +65,11 @@ public class Bid extends BaseDomain{
         this.bidCatalog = bidCatalog;
     }
 
-    public BigDecimal getScore() {
+    public int getScore() {
         return score;
     }
 
-    public void setScore(BigDecimal score) {
+    public void setScore(int score) {
         this.score = score;
     }
